@@ -1,21 +1,9 @@
 package glass.chungwon.verse;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.AppCompatEditText;
-import androidx.appcompat.widget.AppCompatImageButton;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-import androidx.appcompat.widget.Toolbar;
-
-import com.android.volley.toolbox.JsonArrayRequest;
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
-
 import android.os.Parcelable;
 import android.util.Log;
 import android.view.KeyEvent;
@@ -24,29 +12,33 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.inputmethod.EditorInfo;
 import android.view.inputmethod.InputMethodManager;
-import android.text.Html;
-import android.text.Spanned;
-import android.text.method.LinkMovementMethod;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.AppCompatEditText;
+import androidx.appcompat.widget.AppCompatImageButton;
+import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
+import com.android.volley.toolbox.JsonArrayRequest;
 import com.android.volley.toolbox.Volley;
-
-import java.io.UnsupportedEncodingException;
-import java.lang.reflect.Type;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
-import org.json.JSONArray;
-
+import java.io.UnsupportedEncodingException;
+import java.lang.reflect.Type;
 import java.net.URLEncoder;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 
 /**
@@ -68,17 +60,17 @@ public class ItemListActivity extends AppCompatActivity {
     /**
      * An array of verse items. It's set to static as it is needed to restore session.
      */
-    public static ArrayList<VerseItem> ITEMS = new ArrayList<VerseItem>();
+    public static ArrayList<VerseItem> ITEMS = new ArrayList<>();
 
     /**
      * A logging tag for Logcat.
      */
-    private String TAG = "ItemListActivity";
+    private final String TAG = "ItemListActivity";
 
     /**
      * A map of verse items, by ID.
      */
-    public static Map<String, VerseItem> ITEM_MAP = new HashMap<String, VerseItem>();
+    public static Map<String, VerseItem> ITEM_MAP = new HashMap<>();
 
     private SimpleItemRecyclerViewAdapter recyclerViewAdapter;
     private RecyclerView recyclerView;
@@ -115,9 +107,8 @@ public class ItemListActivity extends AppCompatActivity {
             editText.setText(savedInstanceState.getString("queryString"));
             performSearch();
             mLayoutManager.onRestoreInstanceState(state);
-        } else {
-            // Probably initialize members with default values for a new instance
-        }
+        }  // Else { Probably initialize members with default values for a new instance }
+
 
         editText.setImeActionLabel("Search", KeyEvent.KEYCODE_ENTER);
         editText.setOnEditorActionListener((v, actionId, event) -> {
@@ -135,23 +126,18 @@ public class ItemListActivity extends AppCompatActivity {
             return false;
         });
 
-        querySubmitButton.setOnClickListener(new AppCompatImageButton.OnClickListener() {
-            public void onClick(View v) {
-                performSearch();
-                InputMethodManager imm = (InputMethodManager) getSystemService(INPUT_METHOD_SERVICE);
-                imm.hideSoftInputFromWindow(querySubmitButton.getWindowToken(), InputMethodManager.HIDE_NOT_ALWAYS);
-            }
+        querySubmitButton.setOnClickListener(v -> {
+            performSearch();
+            InputMethodManager imm = (InputMethodManager) getSystemService(INPUT_METHOD_SERVICE);
+            imm.hideSoftInputFromWindow(querySubmitButton.getWindowToken(), InputMethodManager.HIDE_NOT_ALWAYS);
         });
 
 
         FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                // TODO: Not implemented, will open new Chrome page with the url of the verse, currently invisible in the UI
-                Snackbar.make(view, "Going to verse URL", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-            }
+        fab.setOnClickListener(view -> {
+            // TODO: Not implemented, will open new Chrome page with the url of the verse, currently invisible in the UI
+            Snackbar.make(view, "Going to verse URL", Snackbar.LENGTH_LONG)
+                    .setAction("Action", null).show();
         });
 
         if (findViewById(R.id.item_detail_container) != null) {
@@ -181,7 +167,7 @@ public class ItemListActivity extends AppCompatActivity {
         editText = (AppCompatEditText) findViewById(R.id.search_text);
         recyclerView = findViewById(R.id.item_list);
 
-        String query = editText.getText().toString();
+        String query = Objects.requireNonNull(editText.getText()).toString();
         Log.d(TAG,"Saved queryString: " + query);
         savedInstanceState.putString("queryString", query);
 
@@ -192,14 +178,14 @@ public class ItemListActivity extends AppCompatActivity {
     public void onPause() {
         super.onPause();
 
-        state = recyclerView.getLayoutManager().onSaveInstanceState();
+        state = Objects.requireNonNull(recyclerView.getLayoutManager()).onSaveInstanceState();
         //read current recyclerview position
         index = mLayoutManager.findFirstVisibleItemPosition();
         View v = recyclerView.getChildAt(0);
         top = (v == null) ? 0 : (v.getTop() - recyclerView.getPaddingTop());
 
         editText = (AppCompatEditText) findViewById(R.id.search_text);
-        queryString = editText.getText().toString();
+        queryString = Objects.requireNonNull(editText.getText()).toString();
     }
     @Override
     public void onResume() {
@@ -211,13 +197,13 @@ public class ItemListActivity extends AppCompatActivity {
         {
             mLayoutManager.scrollToPositionWithOffset( index, top);
         }
-        if(queryString != "") {
+        if(!Objects.equals(queryString, "")) {
             editText = (AppCompatEditText) findViewById(R.id.search_text);
             editText.setText(queryString);
         }
     }
     @Override
-    public void onRestoreInstanceState(Bundle savedInstanceState) {
+    public void onRestoreInstanceState(@NonNull Bundle savedInstanceState) {
         // Always call the superclass so it can restore the view hierarchy
         super.onRestoreInstanceState(savedInstanceState);
         Log.d(TAG,"restoring saved instance");
@@ -228,13 +214,13 @@ public class ItemListActivity extends AppCompatActivity {
         // Restore value of members from saved state
         editText.setText(savedInstanceState.getString("queryString"));
         performSearch();
-        ((LinearLayoutManager) recyclerView.getLayoutManager()).scrollToPositionWithOffset(savedInstanceState.getInt("lastFirstVisiblePosition"),0);
+        ((LinearLayoutManager) Objects.requireNonNull(recyclerView.getLayoutManager())).scrollToPositionWithOffset(savedInstanceState.getInt("lastFirstVisiblePosition"),0);
         savedInstanceState.putInt("lastFirstVisiblePosition", 0);
     }
 
     private void performSearch() {
         editText = (AppCompatEditText) findViewById(R.id.search_text);
-        String query = editText.getText().toString();
+        String query = Objects.requireNonNull(editText.getText()).toString();
         Log.d(TAG, "received search param: " + query);
         String searchText;
         try {
@@ -252,28 +238,17 @@ public class ItemListActivity extends AppCompatActivity {
 
         // Volley request to the url that populates the recycler view
         JsonArrayRequest getRequest = new JsonArrayRequest(Request.Method.GET, url + query, null,
-                new Response.Listener<JSONArray>()
-                {
-                    @Override
-                    public void onResponse(JSONArray response)
-                    {
-                        Type verseListType = new TypeToken<ArrayList<VerseItem>>(){}.getType();
-                        Gson gson = new Gson();
-                        ITEMS = gson.fromJson(String.valueOf(response), verseListType);
-                        for(VerseItem verse : ITEMS) {
-                            ITEM_MAP.put(verse.id, verse);
-                        }
-                        recyclerViewAdapter.swap(ITEMS);
-
-                    }
-                },
-                new Response.ErrorListener()
-                {
-                    @Override
-                    public void onErrorResponse(VolleyError error) {
-                        Log.e(TAG, error.toString());
-                    }
+            response -> {
+                Type verseListType = new TypeToken<ArrayList<VerseItem>>(){}.getType();
+                Gson gson = new Gson();
+                ITEMS = gson.fromJson(String.valueOf(response), verseListType);
+                for(VerseItem verse : ITEMS) {
+                    ITEM_MAP.put(verse.id, verse);
                 }
+                recyclerViewAdapter.swap(ITEMS);
+
+            },
+            error -> Log.e(TAG, error.toString())
         );
 
         //Instantiate the RequestQueue and add the request to the queue
@@ -313,6 +288,7 @@ public class ItemListActivity extends AppCompatActivity {
             }
         };
 
+        @SuppressLint("NotifyDataSetChanged")
         public void swap(List<VerseItem> list){
             if (mValues != null) {
                 mValues.clear();
@@ -332,6 +308,7 @@ public class ItemListActivity extends AppCompatActivity {
             mTwoPane = twoPane;
         }
 
+        @NonNull
         @Override
         public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
             View view = LayoutInflater.from(parent.getContext())
@@ -339,6 +316,7 @@ public class ItemListActivity extends AppCompatActivity {
             return new ViewHolder(view);
         }
 
+        @SuppressLint("SetTextI18n")
         @Override
         public void onBindViewHolder(final ViewHolder holder, int position) {
             holder.mIdView.setText(mValues.get(position).title + "\n" + mValues.get(position).reference);
@@ -353,7 +331,7 @@ public class ItemListActivity extends AppCompatActivity {
             return mValues.size();
         }
 
-        class ViewHolder extends RecyclerView.ViewHolder {
+        static class ViewHolder extends RecyclerView.ViewHolder {
             final TextView mIdView;
             final TextView mContentView;
 
